update stock and stock entry service, update ultilities method

+++ CafeManagement/CafeManagement/Models/StockEntryDetail.cs
@@ -5,9 +5,9 @@
         public Guid Id { get; set; } 
         public Guid StockEntryId { get; set; }
         public Guid IngredientId { get; set; }
-        public decimal Quantity { get; set; }
+        public float Quantity { get; set; }
         public decimal Price { get; set; }
-        public decimal TotalValue => Quantity * Price;
+        public decimal TotalValue => (decimal)Quantity * Price;
 
         public StockEntry StockEntry { get; set; }
         public Ingredient Ingredient { get; set; }
